requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
scikit-learn
dill
pyaml
mlflow
dagshub
fastapi
uvicorn

##-e .